# static/swagger.json
{
  "info": {
    "title": "Midilarity API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/artist_feeling": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "artist",
            "required": true,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get Artist Feeling"
      }
    },
    "/intensity_ratio": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get Intensity and Tension Ratio"
      }
    },
    "/popular_sequence": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get Most Popular Sequence"
      }
    },
    "/positive_artists": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get Positive Artists"
      }
    },
    "/similar": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "required": true,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          },
          {
            "in": "query",
            "name": "artist",
            "required": true,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get Similar Songs"
      }
    },
    "/song_by_feeling": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "feeling",
            "required": true,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get By Feeling"
      }
    },
    "/songs": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          },
          {
            "in": "query",
            "name": "artist",
            "required": false,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get All Songs"
      },
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "artist": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Insert Song"
      }
    },
    "/tracks": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "required": true,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          },
          {
            "in": "query",
            "name": "artist",
            "required": true,
            "schema": {
              "$ref": "#"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Get Song Tracks"
      }
    }
  },
  "tags": []
}